shifting segmentation comparison to proper seg_compare.py file
